Add Animals with (in)visibility functionality
Proof of concept working with emoji. Every Animal component has its own
visibility threshold, and the Game keeps track of how many ticks without
sound violations have occurred. When the number of silent ticks exceeds
an Animal's visibility threshold, the animal is added to a state array
visibleAnimals. When the sound threshold is crossed, the array is
cleared and the silent ticker set back to zero.